[Fix] 슬릭 슬라이더 수정완료
